Refactored projects api route

+++ pages/api/projects/[id].ts
@@ -1,6 +1,7 @@
 import dbConnect from '@util/db/dbConnect';
 import { ProjectModel } from '@models/Project';
 import { NextApiHandler } from 'next';
+import { getSession } from 'next-auth/client';
 
 dbConnect();
 
@@ -10,6 +11,8 @@ const handler: NextApiHandler = async (req, res) => {
     method,
   } = req;
 
+  const session = await getSession({ req });
+
   switch (method) {
     case 'get':
     case 'GET':
@@ -17,26 +20,30 @@ const handler: NextApiHandler = async (req, res) => {
         const project = await ProjectModel.findById(id);
 
         if (!project) {
-          return res.status(400).json({ message: '400, Bad Request' });
+          return res.status(400).json({ message: 'Bad Request' });
         }
 
         return res.status(200).json(project);
       } catch (error) {
-        return res.status(400).json({ message: '400, Bad Request', error });
+        return res.status(400).json({ message: 'Bad Request', error });
       }
       break;
     case 'delete':
     case 'DELETE':
+      if (!session) {
+        return res.status(401).json({ message: 'Unauthorized request' });
+      }
+
       try {
         await ProjectModel.findByIdAndDelete(id);
 
         return res.status(200).send('Project deleted');
       } catch (error) {
-        return res.status(400).json({ message: '400, Bad Request', error });
+        return res.status(400).json({ message: 'Bad Request', error });
       }
       break;
     default:
-      return res.status(400).json({ message: '400, Bad Request' });
+      return res.status(400).json({ message: 'Bad Request' });
       break;
   }
 };
